requirements.txt
rich
kafka-python
hdfs
apache-flink==1.17.1
nltk